requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0

# Google Sheets
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Financial metrics
scipy>=1.11.0
pyxirr>=0.10.0

# PDF parsing (CAMS statements)
pdfplumber>=0.10.0

# Scheduling
APScheduler>=3.10.0

# Anthropic
anthropic>=0.34.0

# Notifications
twilio>=8.10.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
